+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See also [https://github.com/stkb/vscode-rewrap/releases](https://github.com/stk
 
 - Latex: Don't wrap verbatim and math sections (#68, #69)
 - Added Crystal language support
+- VSCode extension now activates on startup
 
 
 ### 1.6.1